Rpi4/requirements.txt
smbus
mpu6050-raspberrypi
azure-iot-device
RPi.GPIO
adafruit-circuitpython-tca9548a

# https://learn.adafruit.com/circuitpython-on-raspberrypi-linux/installing-circuitpython-on-raspberry-pi